i18n: Add basic framework for latin
This mostly works on my VM, but  needs the
Latin locale, which will be added in a later commit.

This does prepare the inclusion of Latin as a language
to be selected, but does not render it yet. To render
it, the language can just be added in the config.