frontend: Add cdedburl workaround for non-unique keys
Before Werkzeug 0.15, MultiDicts are not handled properly
by the build.urls function, leading to all values except
for the last being dropped.

However, it works if instead of multiple keys, one key with
a list of values is passed. Nevertheless, this breaks when
we do magic replacements of <*_id> placeholders in the path.
Since this can only happen for unique keys, this in turn can
be fixed by unwrapping the values of the dict if possible.

After upgrading to Debian Bullseye, this workaround should
be unnecessary and be dropped.

Closes #1701