doc/ml: Update mailinglist management documentation
This updates the mailinglist management documentation to
quite a multitude of changes which have accumulated without
adjusting the documentation. This inculdes the new action
to reset unsubscriptions as well.

Furthermore, it redoes the wording to match the expressions
we use in our code.

Last but not least, it fixes the same issues in the graph and
improves its design substantially.