core: Fix quota for own profile
This fixes the issue in two redundant ways.

First, it puts all QuotaException into an Atomizer as intended.
Therefore, the quota can no longer rise above the threshold.

Second, the mechanism is made more robust by exempting accesses
to the own profile only completely.
Now, we do not only not increment the quota when accessing your
own data, we do not even check it..

This commit allows a user to always access his own profile,
independent of any quotas.

Closes #1760