js: Make cdedbSearchPerson more robust
If null was provided as value for the list of users
to be excluded before, the function did not render
the user dropdown even though it was sent the
correct data from the server.

This change makes the js more robust by parsing
any nullish value for the list of excluded users
to an empty list.