doc: Document enums
This documents all enums which are saved as enums
in the database and can thus be directly user-facing.

Since we display such information only in CSV or JSON
files for automated processing, I have deemed the API
section of our documentation the appropriate place.

Furthermore, it adds some relevant links to this
information.

Closes #917