i18n: po files: improve sorting by file
The merge driver now also sorts the messages, and in the Makefile we sort at the end now.

I do not know if the latter change has an advantage, but with the sometimes obscure gettext functionality it could well be that there is some.
There should definitely be no disadvantage.