i18n: merge driver: remove steps without effect
msgcat also considers messages as unique, which are present in both, but have fuzzy markers in one of both files.
Since the conflicts are marked as fuzzy, the local-only and remote-only files contain also the messages which caused the conflict.
Thus, the local-only and remote-only files are equal to the local-changes and remote-changes, respectively.