core: Use finite list for country names
This changes our country handling to use a finite list
of country names instead of a free text field. Hence, we
no longer get misleading countries like "NRW".

This is important since we use the country information
to send mail to our members. Therefore, it must be in a
form the postal office can understand, with countries
either in German, English or French. To be most universally
understandable and simple, I chose an English list.

There are multiple possible ways to go about localizing this:
* Just do not do it.
* Use (v, gettext(v)) as tuples, with v either being English
  or the country code
* Use another avenue of localization which is easier to maintain,
  like a separate dict.

This can not be deployed as is, since the selection form
will delete the current value if there is no suitable
present. Hence, one would require to either deploy the
validation slightly earlier than the select form, or create
a migration script.