tests: add basic infrastructure for running tests in parallel
This already works with passing the database name as environment variable on the command line, as e.g.

`TESTDATABASENAME=cdb_test_3 bin/singlecheck.sh show_event`

It is only tested for docker yet. Currently, a few tests error during `make sample-data-test-shallow` call.

For getting this feature working in the auto-build, presumably the postgres and pgbouncer files have to be adjusted.
Also, documentation is missing.