assembly: Add link to mailinglist
This actually adds a useful reference to the
Wahlbekanntmachung mailinglist, sucht that the user
can easily subscribe.

We decided to do this to be able to link to
mailinglists as such without having to introduce
new inter-realm dependencies. Furthermore, it
allows in principle to specify a link to a
mailinglist which is not managed by the cdedb.

Closes #1579