ml: Introduce DatabaseSubscriptionState
This is a proposal to make sure we are only
querying SubscriptionStates from the database which
are actually stored there.

This is done by introducing a new validator and using
it inside the respective backend functions.
For simplicity's sake, we are not enforcing the
respective validation type by method signatures,
but only by affirm statements deeper inside.

I would have liked to introduce an enum function
`is_written` to access this information, however this
is sadly not possible due to the impossibility of
enum inheritance.

In principle, this approach could be extended in the
following ways:
* Actually enforce the use of DatabaseSubscriptionStates
  in the function signatures. This would require quite
  excessive casting at callsites.
* Actually pimp up the backend functions to work
  correctly for SubscriptionState.none .