Fix Apache invokation in docker container
Previously we used `apachectl` directly.
This had the drawback that it received the signals from docker
which it did not pass along to the actuall apache process.
Directly invoking `apache2` is problematic
as some environment variables need to be loaded and set.
The workaround is to set an environment variable for `apachectl`
which then execs `apache2` replace itself with it.

Additionally the stopsignal was changed from SIGTERM
to SIGWINCH which performs a graceful shutdown.