common: Be more restrictive regarding database typing
Even though we are not using an ORM, we still have some
guarantees which kind of data we are passing into our
specific database functions.

This commit enforces the corresponding restrictions as
far as respectively possible. The more specific a sql
function, the more restrictive the input types can be.