Shift target state map back to enum
Shifting this map to SubscriptionManager makes the code
less understandable.

Even more problematically, changing the target states of
SubscriptionActions can break implicit assumption in the
remainder of subman code, in particular regarding
unsubscribing actions and _check_policy_requirements.

Last but not least, full customizability was not met
anyway, since it was impossible to add additional actions.