assembly/attachments: extract logic in which way an attachment(_version) may be interacted with
this uses the already existing backend helper functions to check if an attachment version
may be added, changed or deleted.

Also adds a frontend endpoint to delete an attachment, which is only allowed
if it has only one version left to avoid user errors. However, the backend allows
to remove the last attachment version without deleting the attachment in principle