genesis: Allow admins to specify past events for cde genesis
This allows admins to specify past events for cde genesis.
Furthermore, it reworks the ui for the genesis request to show
cde data in more appealing fashion.

For now, this does not include a possibility to specify past courses,
since this is a bit problematic if the past event is not known.
It would be kind of easy to do so via two edits, where you can
only choose a past course if you have specified a past event before,
but since this feature was not requested specifically, I left it out
here.

Closes #2127