validation: handle ignoration of ValidationWarnings magically
This introduces a new magic request value specifying if the warnings
shall be ignored during validaiton in the frontend or not. This is
parsed directly in application.py and stored in rs.ignore_warnings.

This changes validate_check to explicitly ask if warnings shall
be ignored or not. Mostly, this is masked by the check_validation
wrapper.

Additionally, validate_affirm always ignores validation warnings.

see proposal in #2009