backend&frontend: introduce new wrapper verify_validation
this is another thin wrapper around validate_check which returns
all encountered errors instead of appending them to the rs or raising
them. This is needed if the error handling is special, f.e. in money parsing.
Currently, this ignores all warnings.

This eliminates all direct uses of validate_check in our code.