assembly: Add comment for concluded ballots
This adds the possibility of adding a comment for
concluded ballots. This contains the complete backend
for the feature including tests, as well as the display
in the frontend and a frontend endpoint.

The template for actually changing the comment is
still missing though.

For some reason, are_ballots_voting sometimes returns
None, in particular for ballot 15.
This can not be reproduced by performing the query
manually.