assembly: Don't use MAGIC_ABSTAIN in own_vote for preferental voting
This is basically a revert of 08e86f84ec1879b5eacea3a64e631b43135f7204

We need to keep the vote string as-is for filling in the current value
in the preferential voting form. In addition I decided to keep the raw
vote string for the "own vote" section on the show_ballot_result page.

Fixes #2293