coverage: Add further unreachable statements
This adds no cover comments to further statements which
are "untestable", since they are blanket checks where we
would catch the errors earlier if they were known, are
not triggerable in test instances, would require shifting
the whole test environment around, or even are plain
unreachable.

This intentionally does not exclude further spots which
are hard/unconventional to test, but testable.