ldaptor: refactor construction and usage of ldap-tree
This splits the handling of ldap branch entries and ldap leaf entries
more clearly. At the same time, this introduces some redundancy
(we require the keys of the hardcoded dicts to be consistent)
to avoid the need to use iterative dicts.

This has the great advantage that the attributes of mulitple entries
can now be retrieved at the same time, which redruces the number of
database calls to instantiate the children of f.e. the user entry
from n to 1.