test_script: refactor script test to be agnostic about database contents.
Now, it only relies on the tables being present, not on their contents.
As other tests never change tables just their contents, this should be sufficient.

Populating/setting up the database on some higher level in preparation of the script tests
seems not useful, since it would use sql scripts. (Via psql as subprocess they are ugly, and
depending on the script facility we are going to test is obviously wrong.)