event: Rework payment information
Due to some new thoughts from the meeting today,
I adjusted the semantics here to properly show what
is there to pay right now. Therefore, I reverted
the adjustments of Johannes to this very point
where we had not taken this into account yet.

Furthermore, it generally fixes some logical bugs
in behavior if there is a fee, but no iban.

In a second step, one might wonder whether we
should actually allow to make payment information
adjustable once registrations are present. This
would be helpful for MusikAka and similar cases.

Closes #2453
Closes #1956 in another way