core/change_user: add special validation for donaition in combination with lastschrift
On the one hand, the donaition may only lay in a specific value range.
Note that this does not hold if an admin changes the donaition of an users,
and that the validation is also passed if the user does not change the value.

On the other hand, we warn the user explicitly if he change the donaition,
but is not the account owner himself. So, we "ensure" the account owner
gives his consent to the changed donaition.