ldap/entry: prevent database ddos attacks
The basic access restriction is implemented during filtering the results
of a search query. However, since some of the entries are loaded
dynamically from the database, this performs load, even if the requesting
user is not permitted to see the data.

Therefore, we lower the (otherwise) clean separation between data construction
(in entry.py) and security restrictions (in server.py), so we perform only
queries on the server if the requesting user is not
a) Anonymous
b) A persona (unless requesting its own data)

This should hopefully prevent a number of attack szenarios.