ldap/backend: encrypt dua passwords once during startup
Since encrypting the passwords is not that cheap and the dua passwords
can not change at runtime (since they are loaded during startup from
the secrets config), we encrypt them once at startup instead for each
request.

This should heavily increase the speed of the tests, they are now passing
in 30 seconds on my local maschiene, compared to about 600 seconds before.