tests/ldap: add separate ldap test service and use it during tests
This is a bit ugly, since we need to add the whole /cdedb2 module to
the python port to get the relative imports inside the test config
file working.

Additionally, we need to wait for the ldap server to start up
before we can run the tests. This is currently handled by a sleep
in check.py, maybe there is a better way to solve this...