ldap: use psycopg3 to connect to the database
This is another way to circumvent the aiopg-pool - pgbouncer clashing.
The big advantage here is that we can open a single connection on
startup and use it forever, in complete analogy to the syncron database
connections used by the remaining parts of the CdEDB.

This is not possible with aiopg, and it feels more natural to use
the async advantages of psycopg3 directly, instead of wrapping syncron
psycopg2 with aiopg to make it async.