ldap/backend: use psycopg3 pooler for database connections
In the web application, we create a new database connection for each user
during login, and bound the database connection to the users session.
To load-balance this, we use pg_bouncer in the background.

However, we do not have the concept of a 'session' in the ldap backend.
Therefore, we use the psycopg3 pool to "emulate" user session:
The pool gets a connection from pg_bouncer, stores and reuse it.
After a lifetime of 1 hour, it releases the connection and gathers a
new one.

The size of the pool is per default 1 connection, but may go up until 10.
If the traffic lessens, the additional connections will not be refetched
once they expired.

Since we have a limit of 200 connections for pg_bouncer, this seems to
be sensible sizes to start with. We may tweak this if we encounter some
performance issues in one way or the other.