script: Repair lost timestamps in various logs.
The initial migration lost some information by not transplanting timestamps of
the account changelog. We rectify this by retroactively adjusting the
timestamps. This is normally not possible as the logs are write only so a
special role is introduced allowing this. Care is taken to check that the
adjustments target the correct entries. Note that everything happens in one
big transaction making it incompatible with concurrent accesses.

DEPLOY: Graft history by following related/deploy/2022-09-07-history-graft.txt.