common/fields: move get_persona_fields_by_realm back into core_frontend
The function states a more general usecase as it coveres. In fact, the
fields in PERSONA_FIELDS_BY_REALM are rather arbitrary. The helper
function was designed to determine the fields which are modifiable
in (admin_)change_persona, and it should only be used there.

The (wrong) usage of the PERSONA_FIELDS_BY_REALM (which are comulative
instead of repetitive, in contrast to all other field constans defined
in common/fields) was the reason for information loss during genesis
decision, see #2624