docker/cdedb-entrypoint: enhance order of setup steps
This ensures that the database, app and ldap container run in correct dependence:
First the database container starts, then the app container creates
the database users and schemas and at last the ldap container can access it.

The ldap container polls the database container before the app container
finished in set up, but I see no easy way to solve this.
Since we use a pooler for ldap, this is no problem and after the
app container finished the setup, the ldap container gets a connection.

closes #2500