postgres: use DROP DATABASE WITH (FORCE)
this drops databases even if there are open connections. This is
usefull for docker, where the ldap container has a persistent database
connections with prevents the app container from applying sample data.

The broken ldap connection does not terminate the ldap service.
The first database query fails, but after that a valid connection
from the pool will be used, and everything works again.

closes #2614