genesis: Links account to be merged
This does not add the respective doppelganger link for
batch_admissions, since this is a bulk operation.

On the other hand, it happens at genesis that you venture
of in the profile of the user to make a link between their
notes and their past events, for example.
Batch admissions does offer no comparable scenario-

Closes #2899