core: Allow inspect_changes for cde_admin
The review facility is only used for the cde realms,
so it is natural to allow review for cde_admins.

Otherwise, user experience is weird either way, since
they can interact with the pending changes via adminchange
and can see a changelog (where a forbidden links was
shown before).

In the backend, they have already been privileged in
changelog_resolve_change accordingly.