core: Properly show pronouns on profile
This fixed the behavior if pronouns_profile is False:
The pronouns are still visible to the user themselves
and admins, who may edit them.
Additionally, they are visible to orgas, who may select
for them using registration query either way.

Furthermore, pronouns_nametags is no longer shown with
cde view, but only to admins and orgas.