cde/show_semester: change semester btns to btn-default if the step is currently running
We do not want to disable the buttons, since something may go wrong and the worker
stop to work - then, we can simply hit the button again and start a new worker.

However, it seems desirable to add a visual discrimination to the button,
to show that currently something is ongoing.