session: Implement authentication as orga droid using specifically formatted API token.
Every token has a primary id, which can be used to reference the specific token, for example in the event log.

A token is linked to a singular event, but not to any single persona.

The formatted token consists of three parts:
- A prefix, indicating what kind of token this is supposed to be: "CdEDB-Orga-".
- An identifier. This hexstring is used to retrieve the hashed secret. (currently 16 unique hex chars).
- A secret. This hexstring is hashed and compared to the salted hash stored in the database. (currently 48 hex chars).

While the identifier needs to be stored in plaintext, the secret mustn't be. Thus the entire token is only displayed once after creation and never afterward.

Tokens can be deauthorized by deleting the stored hash. Tokens can have an expiration time, after which they can no longer be used.

They have a creation and last access time for logging purposes and a title and a notes field for the user to store additional information about each token.