common/merge_dicts: special case multidict-dict case
Since we replaced some entity ids with dataclasses (especially fields,
f.e. lodge_field), the dict representation of the event does not
contain an int at this point, but a dict representation of the field.

So, we check if this is the case and select only the id of the entity
at this point, since this is basically used to populate the
rs.values dict with default parameter in the templates.