docker: Shift paradigm for dev-container.
Previously this tried to also implement a production-container vs. a container
for anything that is needed during development. The new idea keeps the current
container used for running the db (including tests and everything) and adds a
new container for actual development work inside the container (i.e. human
usability).