core: fix error in donation validation
closes #3338

The design flaw here was that we had a check in the template relying on `donation` being a comparable number, which was tricky do deal with in case of validation errors.

I moved this check into the frontend, checking the actual saved data and not the previous input (as this is what the template-check should depend his statement on – whether the user may change his donation themselve or not).