event: Hide courses if there are no tracks
This hides courses everywhere in the orga view,
if no tracks exist. The philosophy behind this is
that orgas should not be concerned with options that
will never be relevant to them over the lifecycle
of their events, while options suddenly appearing
when creating a course could be confusing.

Specifically, this hides references to courses at
* change_event
* downloads
* field_summary

While it keeps the references in
* the admin-only action panel
* part_summary and change_part

Closes #3422