ldap: Improve performance by fetching less
Previously we always queried each user and all groups.
Now, we lower some simpe LDAP filter directly to SQL WHERE clauses.
Additionally, we skip populating memberOf
if only a subset of attributes are requested
and it is not one of them.
This reduces query time on the sample data by a factor of ~40.