event/track_groups: use dataclass in more places
Uses dataclass for REQUESTDICT fields and adopts the validation to
dataclass.

Additionally, I moved more checks inside the validation, and
subsequently splitted the set_track_groups backend function, since
we never make use of the fact that we could create, change and delete
mulitiple track groups at once.

This is very similar to the refactoring of part groups, which I
did in #4147